Add `WithSpan` wrapper
